+++ app/Livewire/CategoriesDelete.php
@@ -2,9 +2,10 @@
 
 namespace App\Livewire;
 
+use Throwable;
 use Livewire\Component;
 use App\Models\Category;
-use Throwable;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class CategoriesDelete extends Component
 {
@@ -41,12 +42,17 @@ class CategoriesDelete extends Component
     public function destroy()
     {
         try {
+            $this->authorize('delete categories');
+
             $this->category->delete();
 
             redirect()->route('categories')
                 ->with('success', 'The category has been deleted successfully.');
         } catch (\Throwable $th) {
-            // Error here
+            if ($th instanceof AuthorizationException) {
+                redirect()->route('categories')
+                    ->with('danger', 'Unauthorized action.');
+            }
         }
     }
 }
